## What this is

Three strands, in 19 commits on `dev`.

### `io:throttle` is a distributed GCRA

**Breaking.** Throttling was a fixed window with a punitive lockout,
counted across gateways by
[comcount](https://github.com/temich/comcount) through the stash
extension's `count`. The window burst across its own edge, the lockout
needed a live timer per blocked key, and counting cost a comcount
counter, a timer and a Redis round trip **per key** — so throttling by
IP under a scan, the case `condition: status 404` exists for, cost as
many round trips as there were addresses. comcount also has to flush
once per window bucket to guarantee one contribution per participant,
which welded staleness to `interval`: a minute-long budget went on group
information up to a minute old.

A key now holds one theoretical arrival time. An admitted request pushes
it `interval / requests` further out, time drags it back, and a request
is admitted while it stays within `interval` of now — so `requests` is
the burst and `requests / interval` the rate it is repaid at, with no
window edge and no lockout to sit out.

What crosses processes is **debt in milliseconds**, not the arrival
time. A duration carries between processes without their clocks having
to agree — the one clock that matters is Redis', read inside the script
— and it is additive in admissions, so each process reports only its own
increments and Redis keeps the running total. Nothing is divided by a
number of nodes, which is as well: the gateway cannot observe how many
of it there are. Reconciling is **one round trip a tick** for every key
and every quota in the process, on a timer of a tenth of `interval`
clamped to 250ms..2s — independent of `interval`, which is what unwelds
staleness from it. Redis keys expire on their own, so nothing sweeps
them.

Deciding reads nothing but a local map, so `settle` is synchronous now:
where it used to walk the loop binding into a component, the request
path holds no promise at all. The 429 carries `Retry-After`, which a
fixed window could not compute. `count` gives way to `meter` in the
stash extension, and comcount goes with it.

Verified against two processes on one Redis: cold, each admits its full
budget independently (the documented one-tick overshoot); after a single
reconcile both see the group debt and both refuse.

### A pass over what a request, a call and a boot pay for

Nine `perf` commits, each with its own measurement in the commit body.

- **Per request** — `auth` sorted its directives per request on an array
every request shares, and families now arrange once per set; span
options per family per stage, built once per route; a `minimatch` glob
compiled per match (2065 -> 229 ns); the URL parsed twice; content
negotiation run per request over one of a handful of strings (685 -> 7
ns). `io:input`/`io:output` project through a set instead of
`Object.entries -> filter -> fromEntries` per entity. A reply that
carries no version was hashed from `JSON.stringify` and then serialized
again by the encoder; the tag now comes from the encoded body.
- **Per call** — endpoint span options and the receiver's span options
memoized; the introspection edge id kept per pair (733 -> 252 ns); RSQL
criteria kept per `Query` (a two-term parse was 6670 ns, the whole call
is now 65). Core no longer deep-clones every materialized record for a
diff only a transition performs.
- **Elsewhere** — `openspan` opens a span only when something records
it, which a default deployment did not (2.8 us x 7-9 per request);
`storages.mongodb` times the call instead of `monitorCommands`, which
was materializing every document twice.
- **Boot** — a component's version hashed `node_modules` one file at a
time through a stream each and in filesystem order; operations booted
serially; indexes created serially; every schema compiled on its own Ajv
instance (4.3 ms cold against 0.6 warm). A five-component composition:
**263 -> 173 ms**.

### Six fixes and the suite

- **`storages.sql` could not write twice** — `Connection` never had
`update`, and `Storage.store` still branched on `_version === 0` while
an entity now carries the version it is stored with, so new entities
took the update path and existing ones were written a version ahead.
- **A schema default never applied on a remote call** — `input` was set
to null before the request was fitted, and the recipient skips
validation of an authentic request, so it was lost for good.
- **A failed composition's ready probe kept the process alive** —
rollback skips connectors still connecting, and the open socket then
never let go.
- **A custom base image never reached a Dockerfile** — `base` guarded
against `null`, but a composition without `image:` has it undefined.
- **A context annotating a custom storage failed with `Cannot find
module`** — annotations are keyed by dependency id, which for a
connector plugged in by a relative path is its package name.
- **`hashd` returned the stream's end event instead of the hash**, which
only went unnoticed because the entry point is always a directory.

The root feature suite is green again: expectations had drifted from the
runtime (base image, storage checksums, `enumerate` requiring a query,
the composition log), and the kubectl/kind scenarios — the only ones
reaching outside the docker compose stack, and `I have a kube context`
switched the developer's own — are gone along with coverage for commands
that no longer exist.

## Breaking changes

`io:throttle` no longer takes `cooldown`, and its schema now rejects
unknown properties, so a declaration carrying one **fails at boot**
instead of being ignored quietly. Metering earns the budget back rather
than locking a key out, so a lockout is expressed as a rate: `requests:
20, interval: 600` admits a burst of twenty and then one every thirty
seconds, where `requests: 20, interval: 60, cooldown: 600` used to
refuse everything for ten minutes.

The stash aspect's `count` is replaced by `meter`, which takes a batch
of keys and debts and answers what the group has reached. It is
documented in the stash readme; `count` never was.

## Verification

`pr.yaml` only runs on PRs into `dev`, so this one gets no automated
checks. Run locally, as that workflow would:

- `npx eslint .` — clean
- `npm run typecheck` — clean
- `npm run test:unit` — **1378 passed**, 201 suites, 0 failed

That last one was red when this branch was cut, and the batch had broken
it itself: `arrange` joined `DirectiveFamily` with the per-request perf
work and `Directive.test.ts`'s mocks were left without it. Nothing said
so, because the workspace runs `jest-esbuild`, which strips types
without checking them, and `tsc` excludes `*.test.ts` — only the root
run type-checks test files. Fixed in the last commit.

Also green: `features:security` **83/83** (634 steps), `io.throttle` 4/4
including the `Retry-After` assertion, and the root `stash` `Metering`
scenario, which asserts the Lua against a live Redis rather than on a
locally-known number the way `Counting` had to.

+++ connectors/storages.sql/src/.migration/.translate/types.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const types = {
-  integer: () => 'integer',
+  integer: () => 'bigint', // timestamps are stored as epoch milliseconds
   string: () => 'varchar'
 }
